+++ dcli/tool/docker/dcli_cli.dart
@@ -11,6 +11,7 @@ import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:dcli/dcli.dart';
 import 'package:path/path.dart';
+import 'package:pubspec_manager/pubspec_manager.dart';
 
 ///
 /// Starts a docker shell with a full install
@@ -60,7 +61,7 @@ class RunCommand extends Command<void> {
     'docker volume create dcli_scripts'
         .forEach(devNull, stderr: (line) => print(red(line)));
 
-    final pubspec = PubSpec.fromFile(DartProject.self.pathToPubSpec);
+    final pubspec = Pubspec.loadFile(DartProject.self.pathToPubSpec);
     final version = pubspec.version.toString();
 
     'docker run -v dcli_scripts:/home/scripts --network host -it $imageName:$version /bin/bash'
@@ -77,7 +78,7 @@ class BuildCommand extends Command<void> {
 
   @override
   void run() {
-    final pubspec = PubSpec.fromFile(DartProject.self.pathToPubSpec);
+    final pubspec = Pubspec.loadFile(DartProject.self.pathToPubSpec);
     final version = pubspec.version.toString();
     final projectRoot = DartProject.self.pathToProjectRoot;
 
@@ -114,7 +115,7 @@ class PushCommand extends Command<void> {
     // }
     // var version = argResults['version'] as String;
 
-    final pubspec = PubSpec.fromFile(DartProject.self.pathToPubSpec);
+    final pubspec = Pubspec.loadFile(DartProject.self.pathToPubSpec);
     final version = pubspec.version.toString();
     print('Pushing version: $version');
     'sudo docker push $imageName:$version'.run;